Fixed image formatting for camera view
